remove dependency on experimental from wazerotest
